set up ESLint for dev tool
