+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 npm-debug.log
 react-components.md
 build
+.eslintrc.json
