+++ package.json
@@ -7,7 +7,8 @@
     "start": "webpack-dev-server --progress --inline --hot",
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack",
-    "watch": "webpack --watch"
+    "watch": "webpack --watch",
+    "eslint": "eslint"
   },
   "author": "Heather Booker",
   "license": "ISC",
@@ -20,6 +21,11 @@
     "babel-preset-react-hmre": "^1.1.1",
     "babelify": "^7.3.0",
     "css-loader": "^0.23.1",
+    "eslint": "^2.11.1",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-plugin-import": "^1.8.1",
+    "eslint-plugin-jsx-a11y": "^1.3.0",
+    "eslint-plugin-react": "^5.1.1",
     "file-loader": "^0.8.5",
     "html-webpack-plugin": "^2.19.0",
     "node-sass": "^3.7.0",
