+++ src/components/Map.jsx
@@ -1,26 +1,18 @@
 var React = require('react');
+var world = require('../assets/world.svg');
 
-
-//build components
-var Content = React.createClass({
+var MapImg = React.createClass({
   render: function() {
     return (
       <div className='row'>
-        <div className='col-md-offset-2 col-md-8'>
-          <Panel />
+        <div className='col-md-12'>
+          <div className='f-map-img'>
+            <img src={world} />
+          </div>
         </div>
       </div>
     );
   }
 });
 
-var Panel = React.createClass({
-  render: function() {
-    return (
-      <div className='f-panel f-panel-big'>
-      </div>
-    );
-  }
-});
-
-module.exports = Content
+module.exports = MapImg
