+++ src/index.jsx
@@ -4,15 +4,9 @@ var ReactDOM = require('react-dom');
 var Nav = require("./components/Nav.jsx");
 var PanelRight = require("./components/PanelRight.jsx");
 var PanelLeft = require('./components/PanelLeft.jsx');
-var MapPage = require('./components/Map.jsx');
+var MapPage = require('./components/MapPage.jsx');
 //require stylesheet
 require("./stylesheets/main.scss");
-//require imgs
-var earth = require("./assets/earth.svg");
-var people = require("./assets/nationalities.svg");
-var gazebo = require("./assets/mixitup.svg");
-//uri vars
-var countries = '#';
 
 
 var Main = React.createClass({
