+++ src/stylesheets/main.scss
@@ -9,6 +9,7 @@
 @import "partials/right-panel";
 @import "partials/background";
 @import "partials/general";
+@import "partials/map-pg";
 //libs
 @import "../../lib/bootstrap-grid/css/bootstrap.css"
