+++ src/stylesheets/modules/_variables.scss
@@ -1,4 +1,10 @@
-$blue-primary: #20A6e7;
+//colours
+$colour-primary: #20A6e7;
+$colour-text-mild: #999999;
+$colour-disabled: #bcbcbc;
+$colour-map: #cc5252;
+
+
 $font-lite: 400;
 $font-med: 500;
 $font-bold: 600;
