+++ src/stylesheets/partials/_background.scss
@@ -14,6 +14,10 @@ body {
   margin: 40px 0px;
 }
 
+.f-panel-big {
+  padding: 15px;
+}
+
 .f-panel-left {
   @extend .f-panel;
   margin: 40px 40px 40px 0px;
