+++ src/stylesheets/partials/_general.scss
@@ -1,4 +1,4 @@
-.f-nav-logo {
+h1, h2, h3, h4, h5, h6, p, span {
   font-family: $raleway;
 }
