+++ src/stylesheets/partials/_nav.scss
@@ -1,5 +1,5 @@
 .f-nav {
-  background-color: $blue-primary;
+  background-color: $colour-primary;
   min-height: 70px;
   position: relative;
 }
